requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
codeproject-ai-api==0.1
configobj==5.0.8
frozenlist==1.4.1
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
multidict==6.0.5
protobuf==5.27.2
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
text-unidecode==1.3
urllib3==2.2.2
voluptuous==0.15.2
yarl==1.9.4
